feat: EPT-2441: separate weight to file
